Refactor upload logic to separate module for clarity
